refactor: converted pages
